fix: add change birthday method
